requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.61.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
